+++ src/main/java/uk/gov/hmcts/juror/support/sql/v2/generation/dto/User.java
@@ -27,7 +27,14 @@ public class User {
     }
 
     public User addCourt(String court) {
-        courts.add(court);
+        this.courts.add(court);
         return this;
+    }
+
+    public boolean hasRole(Role role) {
+        if(this.roles == null){
+            return false;
+        }
+        return this.roles.contains(role);
     }
 }
