+++ src/main/java/uk/gov/hmcts/juror/support/sql/v2/spring/entity/RequirePncCheckViewEntity.java
@@ -2,6 +2,8 @@ package uk.gov.hmcts.juror.support.sql.v2.spring.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.Getter;
@@ -18,5 +20,6 @@ public class RequirePncCheckViewEntity {
     private String jurorNumber;
 
     @Column(name = "police_check")
+    @Enumerated(EnumType.STRING)
     private PoliceCheck policeCheck;
 }
